Correct the tag filter function
The `-f`/`--tag-filter` can be specified many times, and an OSM object
is only included if it matches *all* defined filter's, i.e. a logical
AND of all filters.

The `-F`/`--tag-filter-func` takes a single ordered list of tag filters,
and includes (with `→T`), or excludes (with `→F`), the OSM object based
on the first filter function which matches.

We want to include OSM objects with `canal=…` or `portage=…`. The
original version would only include objects with both tags. This should
include objects with either.